+++ CHANGELOG.md
@@ -14,6 +14,7 @@ For a full diff see [`2.1.0...main`][2.1.0...main].
 - Renamed `Parsed::fromFrontMatterAndContent()` to `Parsed::create()` ([#397]), by [@localheinz]
 - Extracted `BodyMatter` as a value object ([#398]), by [@localheinz]
 - Started composing `Content` and `Data` into `FrontMatter` ([#409]), by [@localheinz]
+- Renamed `Exception\FrontMatterDoesNotHaveKey` to `Exception\DataDoesNotHaveKey` ([#410]), by [@localheinz]
 
 ## [`2.1.0`][2.1.0]
 
@@ -109,5 +110,6 @@ For a full diff see [`4e97e14...0.1.0`][4e97e14...0.1.0].
 [#400]: https://github.com/ergebnis/front-matter/pull/400
 [#407]: https://github.com/ergebnis/front-matter/pull/407
 [#409]: https://github.com/ergebnis/front-matter/pull/409
+[#410]: https://github.com/ergebnis/front-matter/pull/410
 
 [@localheinz]: https://github.com/localheinz